requirements.txt
# Bleeding edge Django
django
django-cors-headers
django-debug-toolbar

# Bleeding edge Rest Framework
coreapi
djangorestframework
git+https://github.com/NiyazNz/django-rest-framework-simplejwt@jwt_cookie
django-url-filter

# Env
python-decouple

# Database
psycopg2-binary

# Images
pillow
